New directory and small fixes
Added a game directory to house the api, ui, and events files that effect the main gameplay. Improved performence of show hide on form elements. Everything pertaining to show and hide is fully built out for forms as of this commit. Updates will be needed as new html is added. Also created a few functions to reset default screen to make code dryer with the show hides.