Reset game board
Reset game board to reflect lunch and learn styling. Updated the events file in the game directory to have the event handler for the click on the game board. The board now updates with X and O based on clicks. Next need to connect the clicks with the api and keep track andupdate the users games.